requirements.txt
stac-validator
siphon
xarray
pystac
netCDF4
intake
sat-search
intake-stac
requests
urllib3
shapely
tzlocal
pytz
python-dotenv
jsonschema
colorlog
lxml
shutil
re
owslib
boto3
